requirements.txt
Django==4.1.6
django-debug-toolbar==3.2.3
djangorestframework==3.12.4